NSL-5363: Name De-duplication: Re-factor code without changing functions - ready to review functions during testing
