--- app/views/de_duplicates/index.js.erb
@@ -1,6 +1,0 @@
-$('#batch-container').removeClass('hidden');
-$('ul#results-tabset li').removeClass('active');
-if (!$('#batch-container').hasClass('has-data')){
-  $('#batch-container').html("<%= j render(:partial => 'index') %>")
-}
-
