+++ app/views/layouts/application.html.erb
@@ -120,7 +120,7 @@ input[type="checkbox"]:not([title]) {
         <div id="stats-container"
              class="main-body-container no-left-padding col-xs-12 col-sm-12 col-md-12 col-lg-12 hidden">
         </div>
-        <div id="batch-container"
+        <div id="all-dep-container"
              class="main-body-container no-left-padding col-xs-12 col-sm-12 col-md-12 col-lg-12 hidden">
         </div>
         <div id="batch-loader-panel-container"
