+++ app/views/names/de_duplicates/transfer_all_dependents/error.js.erb
@@ -1,0 +1,3 @@
+$('.all-dep-message-container').html('');
+$('.all-dep-error-container').html('');
+$('#all-dep-error-message-container').html('<%= @message %>');
