+++ app/views/names/de_duplicates/transfer_all_dependents/success.js.erb
@@ -1,0 +1,4 @@
+$('.all-dep-message-container').html('');
+$('.all-dep-error-container').html('');
+$('#all-dep-info-message-container').html('<%= @message %>');
+
