--- app/views/names/de_duplication/transfer_all_dependents/error.js.erb
@@ -1,3 +1,0 @@
-$('.batch-message-container').html('');
-$('.batch-error-container').html('');
-$('#batch-de-duplicate-error-message-container').html('<%= @message %>');
