--- app/views/names/de_duplication/transfer_all_dependents/success.js.erb
@@ -1,4 +1,0 @@
-$('.batch-message-container').html('');
-$('.batch-error-container').html('');
-$('#batch-de-duplicate-info-message-container').html('<%= @message %>');
-
