+++ config/history/changes-2026.yml
@@ -1,3 +1,7 @@
+- :date: 22-July-2026
+  :jira_id: '5363'
+  :description: |-
+    Name De-duplication: Re-factor code without changing functions - ready to review functions during testing
 - :date: 21-July-2026
   :jira_id: '5864'
   :description: |-
