+++ config/version.properties
@@ -1,1 +1,1 @@
-appversion=5.1.4.51
+appversion=5.1.4.52
