+++ spec/models/ability_spec.rb
@@ -1472,8 +1472,8 @@ RSpec.describe Ability, type: :model do
       allow(session_user).to receive(:qa?).and_return(true)
     end
 
-    it "allows all actions on de_duplicates" do
-      expect(subject.can?("de_duplicates", :all)).to eq true
+    it "allows all actions on names/de_duplicates" do
+      expect(subject.can?("names/de_duplicates", :all)).to eq true
     end
 
     it "allows all actions on tree_versions" do
